Push generation of the short description into the exotic tile component.
